% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{alleleFreqDiff}
\alias{alleleFreqDiff}
\title{alleleFreqDiff}
\usage{
alleleFreqDiff(vcfR = NULL, pm = NULL, p1 = NULL, p2 = NULL, difference = NULL)
}
\arguments{
\item{vcfR}{vcfR object}

\item{pm}{data.frame containing two columns, "id" and "pop". The ids must match those in the vcfR object, but don't need to be in the same order. Each individual must be assigned to a population}

\item{p1}{(character) name of parental population 1}

\item{p2}{(character) name of parental population 2}

\item{difference}{(numeric) allele frequency difference threshold, must be between 0 and 1}
}
\value{
vcfR object
}
\description{
alleleFreqDiff
}
\examples{
#alleleFreqDiff(vcfR = vcfR, pm = pm, p1 = p1, p2 = p2, difference = 0.5)
}
