% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{assign_unique_vertex_indices}
\alias{assign_unique_vertex_indices}
\title{Assign unique vertex indices to all points}
\usage{
assign_unique_vertex_indices(polygons_df)
}
\arguments{
\item{polygons_df}{data.frame containing multiple polygons
distinguished by group and subgroup}
}
\value{
inpput wit new 'vidx' and 'dupe' columns
}
\description{
This will be used later to de-dupe the input data for RTriangle::triangulate()
}
