% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose-old.R, R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decompose complex polygon into a set of simple polygons}
\usage{
decompose(polygons_df, dedupe_with_polyclip = TRUE)

decompose(polygons_df, dedupe_with_polyclip = TRUE)
}
\arguments{
\item{polygons_df}{polygon data.frame with 'subgroup' indicating primary/hole
hierarchy}

\item{dedupe_with_polyclip}{use polyclip to simplify polygons. Default: FALSE,
i.e. do manual deduplication.}
}
\description{
Preserve holes. Even-odd inside test

Preserve holes. Even-odd inside test
}
\examples{
\dontrun{
# Polygon with a hole
polygons_df <- df <- data.frame(
  x        = c(4, 8, 8, 4,   6, 7, 7, 6),
  y        = c(4, 4, 8, 8,   6, 6, 7, 7),
  group    = c(1, 1, 1, 1,   1, 1, 1, 1),
  subgroup = c(1, 1, 1, 1,   2, 2, 2, 2)
)
triangular::decompose(polygons_df)
}
\dontrun{
# Polygon with a hole
polygons_df <- df <- data.frame(
  x        = c(4, 8, 8, 4,   6, 7, 7, 6),
  y        = c(4, 4, 8, 8,   6, 6, 7, 7),
  group    = c(1, 1, 1, 1,   1, 1, 1, 1),
  subgroup = c(1, 1, 1, 1,   2, 2, 2, 2)
)
triangular::decompose(polygons_df)
}
}
