% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_single}
\alias{decompose_single}
\title{Create the S matrix for RTriangle::pslg()}
\usage{
decompose_single(polygon_df, offset)
}
\arguments{
\item{polygon_df}{data.frame containing a single polygon}

\item{offset}{an index counter to ensure different subgroup polygons will
have different vertex indices.}
}
\description{
pslg = Planar Straight Line Graph object. A collection of vertices and segments.
}
\details{
The 'S' matrix is: A 2-column matrix of segments in which each row is a
segment. Segments are edges whose endpoints are vertices in the PSLG, and
whose presence in any mesh generated from the PSLG is enforced. Each
segment refers to the indices in V of the endpoints of the segment. By
default the segments are not specified (NA), in which case the convex
hull of the vertices are taken to be the segments. Any vertices outside
the region enclosed by the segments are eaten away by the triangulation
algorithm. If the segments do not enclose a region the whole triangulation
may be eaten away.
}
