% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{example.popmap}
\alias{example.popmap}
\title{Example of a popmap}
\format{
An object of class \code{data.frame} with 420 rows and 2 columns.
}
\usage{
example.popmap
}
\description{
An S3 object of class data.frame
}
\examples{
data("example.popmap", package = "triangulaR")

}
\keyword{datasets}
