% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.R
\docType{data}
\name{example.vcfR}
\alias{example.vcfR}
\title{Example of a vcfR object}
\format{
An object of class \code{vcfR} of dimension 1215 x 8 x 421.
}
\usage{
example.vcfR
}
\description{
An S4 object of class vcfR
}
\examples{
data("example.vcfR", package = "triangulaR")

}
\keyword{datasets}
