% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{hybridIndex}
\alias{hybridIndex}
\title{hybridIndex}
\usage{
hybridIndex(vcfR = NULL, pm = NULL, p1 = NULL, p2 = NULL)
}
\arguments{
\item{vcfR}{data in vcfR format}

\item{pm}{data.frame containing two columns, "id" and "pop". The ids must match those in the vcfR object, but don't need to be in the same order. Each individual must be assigned to a population}

\item{p1}{(character) name of parental population 1}

\item{p2}{(character) name of parental population 2}
}
\value{
dataframe containing hybrid indices, heterozygosities, and percent missing data
}
\description{
hybridIndex
}
\examples{
#hybridIndex(vcfR = example.vcfR, pm = example.popmap, p1 = p1, p2 = p2)
}
