% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{missing.plot}
\alias{missing.plot}
\title{missing.plot}
\usage{
missing.plot(
  data = NULL,
  outline = T,
  ind.labels = F,
  cex = 2,
  alpha = 1,
  jitter = 0,
  max.overlaps = 10
)
}
\arguments{
\item{data}{Dataframe returned from hybridIndex function}

\item{outline}{(logical) Whether or not to draw possible triangle space as outline}

\item{ind.labels}{(logical) Whether or not to label each individual on the triangle plot}

\item{cex}{(character) Size of points}

\item{alpha}{(numeric) Transparency of points}

\item{jitter}{(numeric) Amount by which to jitter points on plot (to facilitate visualization)}

\item{max.overlaps}{(numeric) Only necessary if labeling individuals. Increasing this will increase the number of individuals labeled, even if labels overlap.}
}
\value{
ggplot2 object
}
\description{
Color the triangle plot by percent missing data in each sample
}
\examples{
#missing.plot(data = your.data)
}
