% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-in-polygon.R
\name{point_in_polygon}
\alias{point_in_polygon}
\title{Determine number of segment crossings from (x, y) to infinity}
\usage{
point_in_polygon(x, y, poly)
}
\arguments{
\item{x, y}{coordinates}

\item{poly}{data.frame with x and y coordinates of all vertices}
}
\value{
count of segment crossings
}
\description{
Determine number of segment crossings from (x, y) to infinity
}
