% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point-in-polygon.R
\name{point_in_polygons}
\alias{point_in_polygons}
\title{Determine number of segment crossings from (x, y) within all polygons}
\usage{
point_in_polygons(x, y, polygons_list)
}
\arguments{
\item{x, y}{coordinates of point to test}

\item{polygons_list}{list of polygon data.frames}
}
\value{
sum of all crossings over all polygons
}
\description{
Determine number of segment crossings from (x, y) within all polygons
}
