% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{simplify_polygon}
\alias{simplify_polygon}
\title{Simplify a single polygon}
\usage{
simplify_polygon(polygon_df)
}
\arguments{
\item{polygon_df}{data.frame containing a single polygon (x, y)}
}
\description{
Simplify a single polygon
}
