% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose-old.R, R/decompose.R
\name{simplify_polygons}
\alias{simplify_polygons}
\title{Simplify a polygon made up of groups and subgroups}
\usage{
simplify_polygons(polygons_df)

simplify_polygons(polygons_df)
}
\arguments{
\item{polygons_df}{data.frame containing multiple polygons
distinguished by group and subgroup}
}
\description{
Simplify a polygon made up of groups and subgroups

Simplify a polygon made up of groups and subgroups
}
