% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{specFreqDiff}
\alias{specFreqDiff}
\title{specFreqDiff}
\usage{
specFreqDiff(vcfR = NULL, pm = NULL, p1 = NULL, p2 = NULL, difference = 0)
}
\arguments{
\item{vcfR}{data in vcfR format}

\item{pm}{data.frame containing two columns, "id" and "pop". The ids must match those in the vcfR object, but don't need to be in the same order. Each sample must be assigned to a population}

\item{p1}{(character) name of parental population 1}

\item{p2}{(character) name of parental population 2}

\item{difference}{(numeric) allele frequency difference threshold, must be between 0 and 1}
}
\value{
dataframe containing allele frequencies in parental pops and differences between them
}
\description{
Calculate allele frequency difference for loci above difference threshold. For all allele frequency differences, set difference at 0
}
\examples{
#specFreqDiff(vcfR = example.vcfR, pm = example.popmap, p1 = 0, p2 = 20, difference = 1)
}
