% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{triangle.plot}
\alias{triangle.plot}
\title{triangle.plot}
\usage{
triangle.plot(
  data = NULL,
  colors = NULL,
  outline = T,
  cex = 2,
  alpha = 1,
  jitter = 0
)
}
\arguments{
\item{data}{(dataframe) Dataframe returned from hybridIndex function}

\item{colors}{(character) Colors to use for each population. Optional, if not supplied, default colors will be generated}

\item{outline}{(logical) Whether or not to draw possible triangle space as outline}

\item{cex}{(character) Size of points}

\item{alpha}{(numeric) Transparency of points}

\item{jitter}{(numeric) Amount by which to jitter points on plot (to facilitate visualization)}
}
\value{
ggplot2 object
}
\description{
triangle.plot
}
\examples{
triangle.plot(data = your.data, colors = your.colors)
}
