% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decompose complex polygons into triangles}
\usage{
decompose(polygons_df)
}
\arguments{
\item{polygons_df}{polygon data.frame with \code{x,y} coordinates,  'group'
column denoting coordinates which belong to the same group,
and 'subgroup' indicating holes within that polygon.  It is assumed
that any other column values are consistent across the entire group
e.g. 'fill'}
}
\value{
data.frame with \code{x}, \code{y} coordinates of the vertices of the
        resulting triangles. Also includes \code{idx} numbering the triangles.
}
\description{
This function wraps \code{RTriangle::triangulate()} to make it easier to
call for my use case, and returns results appropriate for plotting in
\code{ggplot2}.
}
\details{
By combining \code{polyclip} and \code{RTriangle} packages, this package
will successfully decompose into triangles the following polygon types:

\itemize{
\item{Simple polygons}
\item{Polyons with holes}
\item{Multiple polygons with holes}
\item{Polygons with self-intersection}
\item{Polygons with duplicated vertices}
}
}
\examples{
\dontrun{
# Polygon with a hole
polygons_df <- df <- data.frame(
  x        = c(4, 8, 8, 4,   6, 7, 7, 6),
  y        = c(4, 4, 8, 8,   6, 6, 7, 7),
  group    = c(1, 1, 1, 1,   1, 1, 1, 1),
  subgroup = c(1, 1, 1, 1,   2, 2, 2, 2)
)
triangles_df <- triangular::decompose(polygons_df)
ggplot(triangles_df) +
    geom_polygon(aes(x, y, fill = as.factor(idx)))
}
}
