% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose}
\alias{decompose}
\title{Decompose complex polygon into a set of simple polygons}
\usage{
decompose(polys_df)
}
\arguments{
\item{polys_df}{polygon data.frame with 'subgroup' indicating primary/hole
hierarchy}
}
\description{
Preserve holes. Even-odd inside test
}
\examples{
\dontrun{
# Polygon with a hole
polys_df <- df <- data.frame(
  x        = c(4, 8, 8, 4,   6, 7, 7, 6),
  y        = c(4, 4, 8, 8,   6, 6, 7, 7),
  group    = c(1, 1, 1, 1,   1, 1, 1, 1),
  subgroup = c(1, 1, 1, 1,   2, 2, 2, 2)
)
triangular::decompose(polys_df)
}
}
