% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose.R
\name{decompose_single}
\alias{decompose_single}
\title{Inner function to decompose a single polygon 'group'}
\usage{
decompose_single(polygons_df)
}
\arguments{
\item{polygons_df}{polygon data.frame with \code{x,y} coordinates,  'group'
column denoting coordinates which belong to the same group,
and 'subgroup' indicating holes within that polygon.  It is assumed
that any other column values are consistent across the entire group
e.g. 'fill'}
}
\description{
Inner function to decompose a single polygon 'group'
}
