% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{missing.plot}
\alias{missing.plot}
\title{Color the triangle plot by perecent missing data in each sample}
\usage{
missing.plot(data = NULL, outline = T, cex = 2, alpha = 1, jitter = 0)
}
\arguments{
\item{data}{Dataframe returned from hybridIndex function}

\item{outline}{(logical) Whether or not to draw possible triangle space as outline}

\item{cex}{(character) Size of points}

\item{alpha}{(numeric) Transparency of points}

\item{jitter}{(numeric) Amount by which to jitter points on plot (to facilitate visualization)}
}
\value{
ggplot2 object
}
\description{
Color the triangle plot by perecent missing data in each sample
}
\examples{
#missing.plot(data = your.data)
}
